1. Coding standards: a. use space instead of tabulation b.  let's add extra rule: 'insert space between "if", "else" "else if" keywords and boolean condition’
2. JS:
a. Moving JS library to vendor/assets/javascripts and removing duplicate entries.
b. I cannot find a reference in JQuery docs to suggest it is deprecated. But, I agree “JQuery.on” would be the better option. - done.
c. JQuery selector caching. Used $_this instead of $this. ( because of the coding standards mention in airbnb js standards. ( should I revert ?)